Refactor the sample_velocity_of_target method out of ElastivSVT. Needed for latter use in DBRC.
