+++ CMakeLists.txt
@@ -42,6 +42,7 @@ set(PNDL_SOURCE_LIST src/element.cpp
                      src/tabular_energy_angle.cpp
                      src/multiple_distribution.cpp
                      src/cm_distribution.cpp
+                     src/svt.cpp
                      src/elastic_svt.cpp
                      src/energy_grid.cpp
                      src/cross_section.cpp
