+++ src/elastic_svt.cpp
@@ -27,6 +27,8 @@
 #include <functional>
 
 #include "constants.hpp"
+#include "svt.hpp"
+#include "vector.hpp"
 
 namespace pndl {
 
@@ -55,34 +57,6 @@ ElasticSVT::ElasticSVT(const AngleDistribution& angle, double awr,
 
 double ElasticSVT::temperature() const { return kT_ * MEV_TO_EV * EV_TO_K; }
 
-inline double ElasticSVT::Vector::magnitude() const {
-  return std::sqrt(x * x + y * y + z * z);
-}
-
-inline ElasticSVT::Vector ElasticSVT::Vector::rotate(double mu,
-                                                     double phi) const {
-  double xo, yo, zo;
-  const double c = std::cos(phi);
-  const double s = std::sin(phi);
-  const double C = std::sqrt(1. - mu * mu);
-
-  if (std::abs(1. - z * z) > 1.E-10) {
-    const double denom = std::sqrt(1. - z * z);
-
-    xo = x * mu + C * (c * x * z - s * y) / denom;
-    yo = y * mu + C * (c * y * z + s * x) / denom;
-    zo = z * mu - c * C * denom;
-  } else {
-    const double denom = std::sqrt(1. - y * y);
-
-    xo = x * mu + C * (c * x * y + s * z) / denom;
-    yo = y * mu - c * C * denom;
-    zo = z * mu + C * (c * y * z - s * x) / denom;
-  }
-
-  return {xo, yo, zo};
-}
-
 AngleEnergyPacket ElasticSVT::sample_angle_energy(
     double E_in, std::function<double()> rng) const {
   // Direction in
@@ -92,7 +66,10 @@ AngleEnergyPacket ElasticSVT::sample_angle_energy(
   const Vector v_n = u_n * std::sqrt(E_in);
 
   // Get the "velocity" of the target nuclide
-  const Vector v_t = sample_target_velocity(E_in, rng);
+  bool TAR = false;
+  if (E_in >= tar_threshold_ * kT_ && awr_ > 1.) TAR = true;
+  const Vector v_t =
+      TAR ? Vector(0., 0., 0.) : sample_target_velocity(E_in, kT_, awr_, rng);
 
   // Calculate the "velocity" of the center of mass.
   const Vector v_cm = (v_n + (v_t * awr_)) / (awr_ + 1.);
@@ -128,48 +105,6 @@ AngleEnergyPacket ElasticSVT::sample_angle_energy(
   const double mu_lab = u_n.dot(u_n_out);
 
   return {mu_lab, E_out};
-}
-
-ElasticSVT::Vector ElasticSVT::sample_target_velocity(
-    double Ein, std::function<double()> rng) const {
-  if (Ein >= tar_threshold_ * kT_ && awr_ > 1.) {
-    return {0., 0., 0.};
-  }
-
-  const double y = std::sqrt(awr_ * Ein / kT_);
-  double x_sqrd = 0.;
-  double mu = 0.;
-
-  const double P_C49 = 2. / (std::sqrt(PI) * y + 2.);
-
-  bool sample_velocity = true;
-  while (sample_velocity) {
-    if (rng() < P_C49) {
-      // Sample x from the distribution C49 in MC sampler
-      x_sqrd = -std::log(rng() * rng());
-    } else {
-      // Sample x from the distribution C61 in MC sampler
-      const double c = std::cos(PI / 2.0 * rng());
-      x_sqrd = -std::log(rng()) - std::log(rng()) * c * c;
-    }
-
-    const double x = std::sqrt(x_sqrd);
-    mu = 2. * rng() - 1.;
-    const double P_accept =
-        std::sqrt(y * y + x_sqrd - 2. * y * x * mu) / (x + y);
-
-    if (rng() < P_accept) sample_velocity = false;
-  }
-
-  // Get speed of target
-  double s_t = std::sqrt(x_sqrd * kT_ / awr_);
-
-  // Use mu to get the direction vector of the target. We know in the sample
-  // method that we always assume the same incident neutron vector:
-  const Vector u_n{0., 0., 1.};
-  const Vector u_t = u_n.rotate(mu, 2. * PI * rng());
-
-  return u_t * s_t;
 }
 
 }  // namespace pndl
